+++ graphix/sim/statevec.py
@@ -249,7 +249,7 @@ class AbstractStatevector(DenseState, Generic[_ScalarT]):
         >>> circuit = Circuit(2)
         >>> circuit.h(0)
         >>> circuit.cz(0, 1)
-        >>> print(circuit.simulate_statevector().statevec.draw())
+        >>> print(circuit.simulate().state.draw())
         sqrt(2)/2(|00> + |01>)
         """
         return statevec_to_str(