+++ graphix/transpiler.py
@@ -84,22 +84,14 @@ class SimulateResult(Generic[_DenseStateT]):
     """
     Result of a circuit simulation.
 
-    statevec : _DenseStateT
+    state : _DenseStateT
         State representation of the simulation output.
     classical_measures : tuple[int,...]
         Results of classical measurements.
     """
 
-    statevec: _DenseStateT  # mypy rejects covariant types as dataclass parameters as of Python 3.13
+    state: _DenseStateT  # mypy rejects covariant types as dataclass parameters as of Python 3.13
     classical_measures: tuple[int, ...]
-
-
-def _check_target(out: Sequence[int | None], index: int) -> int:
-    target = out[index]
-    if target is None:
-        msg = f"Qubit {index} has already been measured."
-        raise ValueError(msg)
-    return target
 
 
 @dataclass
@@ -526,7 +518,7 @@ class Circuit(InplaceParameterizable):
         return TranspiledPattern(result.pattern, classical_outputs)
 
     @overload
-    def simulate_statevector(
+    def simulate(
         self,
         backend: StatevectorBackend | Literal["statevector"] = ...,
         input_state: Data | None = None,
@@ -537,7 +529,7 @@ class Circuit(InplaceParameterizable):
     ) -> SimulateResult[Statevector]: ...
 
     @overload
-    def simulate_statevector(
+    def simulate(
         self,
         backend: DensityMatrixBackend | Literal["densitymatrix"],
         input_state: Data | None = None,
@@ -548,7 +540,7 @@ class Circuit(InplaceParameterizable):
     ) -> SimulateResult[DensityMatrix]: ...
 
     @overload
-    def simulate_statevector(
+    def simulate(
         self,
         backend: DenseStateBackend[_DenseStateT],
         input_state: Data | None = None,
@@ -558,7 +550,7 @@ class Circuit(InplaceParameterizable):
         stacklevel: int = 1,
     ) -> SimulateResult[_DenseStateT]: ...
 
-    def simulate_statevector(
+    def simulate(
         self,
         backend: DenseStateBackend[_DenseStateT] | _DenseStateBackendLiteral = "statevector",
         input_state: Data | None = None,
@@ -650,7 +642,7 @@ class Circuit(InplaceParameterizable):
         return SimulateResult(_backend.state, tuple(classical_measures))
 
     def visit(self, visitor: InstructionVisitor, *, copy: bool = False) -> Circuit:
-        """Apply `visitor` to all instructions in the circuit.
+        """Apply ``visitor`` to all instructions in the circuit.
 
         Parameters
         ----------
@@ -677,7 +669,7 @@ class Circuit(InplaceParameterizable):
         return self
 
     def apply_angle(self, f: Callable[[ParameterizedAngle], ParameterizedAngle], *, copy: bool = False) -> Circuit:
-        """Apply `f` to all angles that occur in the circuit.
+        """Apply ``f`` to all angles that occur in the circuit.
 
         Parameters
         ----------
@@ -698,12 +690,12 @@ class Circuit(InplaceParameterizable):
 
     def is_parameterized(self) -> bool:
         """
-        Return `True` if there is at least one measurement angle that is not just an instance of `SupportsFloat`.
+        Return ``True`` if there is at least one measurement angle that is not just an instance of :class:`SupportsFloat`.
 
         A parameterized circuit is a circuit where at least one
         measurement angle is an expression that is not a number,
-        typically an instance of `sympy.Expr` (but we don't force to
-        choose `sympy` here).
+        typically an instance of :class:`sympy.Expr` (but we don't force to
+        choose ``sympy`` here).
 
         """
         for instr in self.instruction: