+++ tests/test_clifford.py
@@ -104,6 +104,6 @@ class TestClifford:
         pattern = og.to_pattern()
         pattern_ref = Pattern(input_nodes=[0], cmds=[Command.C(0, c)])
         input_state = rand_state_vector(nqubits=1, rng=fx_rng)
-        state = pattern.simulate_pattern(input_state=input_state, rng=fx_rng)
-        state_ref = pattern_ref.simulate_pattern(input_state=input_state, rng=fx_rng)
+        state = pattern.simulate(input_state=input_state, rng=fx_rng)
+        state_ref = pattern_ref.simulate(input_state=input_state, rng=fx_rng)
         assert state.isclose(state_ref)